Trace: Also track constraints on the current path
Previously, the Trace would only be used to iteratively create an
execution tree where each node is a branch condition. This execution
tree would then be used as the basis for the DSE exploration engine.
That is, a random node in the tree would be picked and negated, the
resulting query would afterwards be used to find an assignment with the
KLEE-based SMT solver which would result in the discovery of a new path.

Apart from the execution tree, the Trace now also tracks the constraints
on the currently executed path through the program. This is necessary as
KLEE tracks constraints independent of expressions. For this reason,
symbolic values where previously evaluate independent of the path
constraints by the evalValue/eval methods which were passed a BitVector
instead of a constrained Query. In the symex-vp context, this only
caused issues with the RegFile::show() Method since this is the only
place where symbolic values are evaluated directly. For concrete values
this is not an issue since they are folded constant expressions,
however, for clarity the Solver method for converting these concrete
values to C types has been renamed to getValue.

To resolve this issue in clover, the evalValue and eval methods
receiving only a BitVector have been removed in favor of variants
accepting a constrained Query as a function parameter. Such a query can
now be generated using the Trace::getQuery method. Since the Trace
tracks constraints on the current path now, it can generated a
constrained query for a given BitVector based on the constraints of the
current path through the program.