Remove functions for evaluation of symbolic values
Since clover implements concolic testing, we only really need to
deal with symbolic values in the tracer (i.e. when we negate a
branch node in the execution tree). There is no need to directly
evaluate a symbolic value to a concrete value otherwise, just use
the concrete part of the Concolic class instead.